pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gui-agent"
version = "0.1.0"
description = "基于视觉的 GUI 自动化框架"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyautogui>=0.9.54",
    "Pillow>=10.0.0",
    "pyperclip>=1.8.2",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["gui_agent*"]
